// src/types/QuinaData.ts
export interface QuinaData {
    dataApuracao: string;
    dataProximoConcurso: string;
    numero: number;
    numeroConcursoAnterior: number;
    numeroConcursoProximo: number;
    dezenasSorteadasOrdemSorteio: string[];
    listaDezenas: string[];
    listaRateioPremio: string[];
    acumulado: boolean;
    valorEstimadoProximoConcurso: number;

}